Add files via upload
Data was collected from a Muse mobile EEG. This data consists of trials of eyes open for 30 sec and trials of eyes closed for 30 sec. Two trials of three were done for each. The files  are labeled "eegData_eyesClosedTrial1_1" where the first number denotes the set and the second denotes the rep.